refactor: extract doltlite_checkout.c
Move checkout/connect-branch implementation and related helpers out of
doltlite_branch.c. Keep dolt_branch command and registration in
doltlite_branch.c; export checkout/connect SQL functions via
doltlite_branch_int.h. Wire build, EXTRA_TSRC, amalgamation, and lint.
