+++ test/lint_layers.sh
@@ -41,6 +41,15 @@ else
   fi
 fi
 
+if [ ! -f "$SRCDIR/doltlite_checkout.c" ]; then
+  lint "$SRCDIR/doltlite_checkout.c: missing — doltlite checkout must stay split"
+else
+  nline=$(wc -l < "$SRCDIR/doltlite_checkout.c" | tr -d ' ')
+  if [ "$nline" -gt 1500 ]; then
+    lint "$SRCDIR/doltlite_checkout.c:$nline lines — doltlite_checkout.c must stay at or below 1500 lines"
+  fi
+fi
+
 
 # Keep doltlite merge split into reviewable units. The core catalog/rows/
 # schema/pass1/pass2 modules must exist (so a re-monolith can't drop a file
